Add files via upload
Building a classifier that recognizes different types of physical activity from signals measured by the accelerometer and gyroscope in your smartphone, which both measure aspects of movement and orientation. The data for this were collected in a lab using a basic smartphone in experiments with human participants carrying out various daily activities in set order.